requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2.10
django-allauth==0.50.0
django-cors-headers==3.11.0
django-filter==21.1
django-ipware==4.0.2
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djangotoolbox==1.8.0
djongo==1.3.6
djoser==2.1.0
dnspython==2.2.1
drf-nested-routers==0.93.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
mongoengine==0.24.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.3
pyparsing==3.0.8
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.2.4
uritemplate==4.1.1
urllib3==1.26.9
